refactor(api): extract registry handlers from system.rs (#3749 10/N)
